+++ CHANGELOG.md
@@ -36,7 +36,10 @@
 * Sort user permissions in alphabetical order on create/edit screen. Refs UIU-1353.
 * Fix memory leak in `UserAccounts.js` discovered by @mkuklis.
 * UI updates to the Permissions modal. Refs UIU-1466.
+* Add permission and permission check for mark a claimed item missing. Refs UIU-1268.
 * Prevent the renewal of claimed returned items. Refs UIU-1261.
+* Adds `checkedInFoundByLibrary` and `checkedInReturnedByPatron` to loan action map. Fixes UIU-1643.
+* Display the correct link for create/reset password. Refs UIU-1608.
 
 ## [3.0.0](https://github.com/folio-org/ui-users/tree/v3.0.0) (2020-03-17)
 [Full Changelog](https://github.com/folio-org/ui-users/compare/v2.26.0...v3.0.0)